# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Radiant Norma API
  description: 'API REST do Radiant Norma — backend Go para validação CADOCs 3040,

    monitoramento regulatório (radar), envios STA e auditoria LGPD/SOC2.


    ## Autenticação (Sprint 7a / v1.6.0)


    Production: JWT bearer RS256 via header `Authorization: Bearer <token>`.

    Issuer pin (`iss`) + `kid` para key rotation. Role-based access (if/admin/readonly).


    Dev mode: `RADIANT_DEV_AUTH=1` aceita `X-IF-ID` como fallback (migration).


    ## Conventions


    - Datas: ISO 8601 (YYYY-MM-DD) ou RFC3339 (com tempo)

    - IDs: integer for radar alerts; string for cadoc codes

    - Errors: JSON `{"error": "<message>"}` com UserError pattern

    - Audit: writes tamper-evident hash chain para todo mutante

    '
  version: 3.0.0
  contact:
    name: Fortvna Radiant Norma
    url: https://fortvna.com.br
servers:
- url: http://localhost:8080
  description: Local development
- url: https://api.radiant-norma.fortvna.com.br
  description: Production (placeholder)
...
paths:
...
  /v1/validate:
    post:
      tags:
      - validate
      summary: Validate CADOC submission
      description: 'L1 parse + L2 semantic (regra) + Optional L3 cross-doc.

        XML ou JSON payload. Returns errors and warnings.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
      responses:
        '200':
          description: Validation result (can have errors)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/sta/submit:
    post:
      tags:
      - sta
      summary: Submit to STA (stub)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Submission'
      responses:
        '200':
          description: Submission accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
...
  /v1/crossdoc/validate:
    post:
      tags:
      - crossdoc
      summary: Cross-document validation (L3)
      description: Validate ecosystem across multiple CADOCs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrossDocRequest'
      responses:
        '200':
          description: Cross-doc validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossDocResponse'
  /v1/auth/dev-token:
    post:
      tags:
      - meta
      summary: Dev-token issuance (Sprint 8a v2.1.0)
      description: 'Dev-only JWT RS256 mint endpoint.


        Em prod: retorna 404 (esconde endpoint). Tokens são emitidos

        por IdP externo (Keycloak/Okta/etc) — Sprint 9+.


        Em dev: ativa via `RADIANT_DEV_TOKEN=1` no backend. Requer

        `RADIANT_DEV_JWT_PRIVATE_KEY` (path) ou

        `RADIANT_DEV_JWT_PRIVATE_KEY_PEM` (inline). Mesma chave pública

        carregada em `RADIANT_JWT_PUBLIC_KEY` aceita os tokens.


        Frontend usa este endpoint para obter JWT real para fluxo

        autenticado. Cookie `rn_jwt` httpOnly é setado pela Next route

        `/api/login` que chama este endpoint.

        '
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevTokenRequest'
      responses:
        '200':
          description: JWT minted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevTokenResponse'
        '400':
          description: Bad request (if_id missing, role inválido, ttl inválido)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Endpoint disabled (RADIANT_DEV_TOKEN not set)
        '503':
          description: Endpoint enabled but signer not configured
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT RS256 with claims {sub, if_id, role, iss, exp, iat}.

        Header: Authorization: Bearer <token>.

        Default kid format: <issuer-prefix>-<key-version>.

        For dev convenience, X-IF-ID header works when RADIANT_DEV_AUTH=1.

        '
  schemas:
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Public-safe error message
...
    ValidationRequest:
      type: object
      required:
      - cadoc_code
      - data_base
      properties:
        cadoc_code:
          type: string
          example: '3040'
        xml:
          type: string
          description: Base64 ou raw XML/JSON
        content_type:
          type: string
          default: application/xml
        data_base:
          type: string
          description: 'Phase 1.6: REQUIRED. Reference date (YYYY-MM or YYYY-MM-DD)'
        versao_layout:
          type: string
          description: Layout version (optional)
    ValidationResponse:
      type: object
      properties:
        passed:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        rules_run:
          type: array
          items:
            type: string
        rules_skip:
          type: array
          items:
            type: string
        duration_ms:
          type: integer
        executed_at:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        code:
          type: string
        codigo:
          type: string
        severity:
          type: string
          enum:
          - E
          - A
        message:
          type: string
          description: Public-safe error message
    Submission:
      type: object
      properties:
        cnpj:
          type: string
        cadoc_code:
          type: string
        data_base:
          type: string
          description: YYYY-MM or YYYY-MM-DD
        xml:
          type: string
        zip:
          type: string
          format: byte
      required:
      - cadoc_code
    SubmissionResponse:
      type: object
      properties:
        protocol_sta:
          type: string
        accepted:
          type: boolean
        rejection:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
        envio_id:
          type: string
        dedup:
          type: string
          enum:
          - idempotency_key
          - xml_hash
          description: 'Phase 4: present when dedup applied'
        warning:
          type: string
...
    CrossDocRequest:
      type: object
      properties:
        cadocs:
          type: object
          additionalProperties:
            type: string
          example:
            '3040': <xml>...</xml>
            '4111': <xml>...</xml>
    CrossDocResponse:
      type: object
      properties:
        passed:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        rules_run:
          type: array
          items:
            type: string
        rules_skip:
          type: array
          items:
            type: string
        duration_ms:
          type: integer
        executed_at:
          type: string
          format: date-time
    DevTokenRequest:
      type: object
      required:
      - if_id
      properties:
        if_id:
          type: string
          maxLength: 64
          description: Tenant identifier (alfanumérico + dash + underscore).
          example: demo-bank
        role:
          type: string
          enum:
          - if
          - admin
          - readonly
          default: if
          description: Role para autorização coarse-grained.
        ttl_seconds:
          type: integer
          minimum: 1
          maximum: 2592000
          default: 86400
          description: 'Token TTL em segundos. Default 24h. Max 30 dias (TTLCap)

            para evitar tokens de vida excessiva acidental.

            '
    DevTokenResponse:
      type: object
      required:
      - token
      - if_id
      - role
      - expires_at
      - ttl_seconds
      properties:
        token:
          type: string
          description: JWT RS256 string. Frontend armazena em cookie httpOnly.
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6ImsxIn0.eyJ...
        if_id:
          type: string
          description: Echo do if_id (sanitizado).
        role:
          type: string
          enum:
          - if
          - admin
          - readonly
        expires_at:
          type: string
          format: date-time
          description: RFC3339 timestamp de expiração.
        ttl_seconds:
          type: integer
          description: TTL aplicado (pode diferir do pedido se clamped).
...
  responses:
...
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'